draw and text functions moved
